Hack in feedback bug reports
This is a change from using the bug_reports app because smashing that
into the feedback form seems hard.

Rather, let's just make a new category for feedback and let admin users
filter on it.

I think this does the same thing, but a little quicker and easier than a
dedicated app.